+++ docker/analyzer.template.entrypoint.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-
-env $ENTRY_POINT_ENV \
-    "$JAVA_17_HOME"/bin/java \
-      -Dorg.seqra.ir.impl.storage.defaultBatchSize=2000 \
-      -Djdk.util.jar.enableMultiRelease=false \
-      -Xmx8g \
-      -jar $ANALYZER_JAR_NAME "$@"